WIP. Implementing my_tx_history_v2_rpc.
