+++ mm2src/coins_activation/src/bch_with_tokens_activation.rs
@@ -2,12 +2,13 @@ use crate::platform_coin_with_tokens::*;
 use crate::prelude::*;
 use crate::slp_token_activation::SlpActivationRequest;
 use async_trait::async_trait;
+use coins::my_tx_history_v2::TxHistoryStorage;
 use coins::utxo::bch::{bch_coin_from_conf_and_params, BchActivationRequest, BchCoin, CashAddrPrefix};
 use coins::utxo::bch_and_slp_tx_history::bch_and_slp_history_loop;
 use coins::utxo::rpc_clients::UtxoRpcError;
 use coins::utxo::slp::{SlpProtocolConf, SlpToken};
 use coins::utxo::UtxoCommonOps;
-use coins::{CoinBalance, CoinProtocol, MarketCoinOps, MmCoin, TxHistoryStorage};
+use coins::{CoinBalance, CoinProtocol, MarketCoinOps, MmCoin};
 use common::executor::spawn;
 use common::mm_ctx::MmArc;
 use common::mm_error::prelude::*;
